rename DAG folder to IR
