+++ src/CMakeLists.txt
@@ -5,4 +5,4 @@ add_subdirectory(Execution)
 add_subdirectory(Conversion)
 add_subdirectory(Diagnostics)
 add_subdirectory(Transforms)
-add_subdirectory(DAG)
+add_subdirectory(IR)
