+++ test/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_subdirectory(TestTools)
 add_subdirectory(AST)
-add_subdirectory(DAG)
+add_subdirectory(IR)
 add_subdirectory(Integration)
 add_subdirectory(Passes)
