+++ test/Integration/CMakeLists.txt
@@ -14,7 +14,7 @@ target_sources(TestIntegration
         Test_ObjectFile.cpp
 )
 
-target_link_libraries(TestIntegration AST Compiler Execution DAG)
+target_link_libraries(TestIntegration AST Compiler Execution IR)
 
 find_package(Catch2)
 target_link_libraries(TestIntegration Catch2::Catch2)
